+++ src/index.ts
@@ -2,3 +2,4 @@ export * from './types';
 export * from './utils';
 export * from './repositories';
 export * from './services';
+export * from './api';