+++ src/repositories/index.ts
@@ -4,3 +4,4 @@ export * from './network.repository';
 export * from './beacon.repository';
 export * from './nfc-tag.repository';
 export * from './policy.repository';
+export * from './leave.repository';